Add model-independent distributed inference workflows (#733)
* Add model-independent distributed inference workflows

* Ignore local review loop state

* Address distributed inference review findings

* Ignore personal worktree helper

* Address final distributed inference review findings

* Harden SGLang export contracts

+++ benchmark/snn_llm/_causal_lm.py
@@ -66,7 +66,10 @@ def forward_step(
             token_count,
             {
                 "lm_loss": torch.stack(
-                    (loss_sum.detach(), token_count.detach().to(loss_sum.dtype))
+                    (
+                        loss_sum.detach().to(torch.float64),
+                        token_count.detach().to(torch.float64),
+                    )
                 )
             },
         )
